requirements.txt
speechrecognition
pyttsx3
spotipy
requests
google-generativeai
